Dirty hack to get around folder names
I don't know why and I should not have to ask why.
But the Arduino IDE insist that it needs to be in the ProtoTracer folder.

However if I make a copy called src.ino (like the folder) and I create another ino file with the actual proper name (ProtoTracer.ino), it just opens the filename based on the folder anyway??

I really hope this don't mess with compiles.